move header info queries
