+++ Frontend-v1-Original/components/header/ConnectButton.tsx
@@ -1,22 +1,18 @@
 import { useAccount, useNetwork } from "wagmi";
-import { useQuery } from "@tanstack/react-query";
 import { Button, Typography } from "@mui/material";
 import {
   ConnectButton as RainbowKitConnectButton,
   useAccountModal,
 } from "@rainbow-me/rainbowkit";
 
-import { QUERY_KEYS } from "../../stores/constants/constants";
+import { useDomain } from "./lib/queries";
 
 export function ConnectButton() {
   const { address } = useAccount();
   const { chain } = useNetwork();
   const { openAccountModal } = useAccountModal();
 
-  const { data: domain } = useQuery({
-    queryKey: [QUERY_KEYS.DOMAIN, address],
-    queryFn: () => resolveUnstoppableDomain(address),
-  });
+  const { data: domain } = useDomain(address);
 
   return address && !chain?.unsupported && domain ? (
     <Button
@@ -34,16 +30,3 @@ export function ConnectButton() {
     <RainbowKitConnectButton showBalance={false} accountStatus="address" />
   );
 }
-
-const resolveUnstoppableDomain = async (address: `0x${string}` | undefined) => {
-  if (!address) return null;
-  const res = await fetch("/api/u-domains", {
-    method: "POST",
-    body: JSON.stringify({
-      address,
-    }),
-  });
-  const resJson = (await res.json()) as { domain: string };
-  if (!resJson?.domain || resJson?.domain === "") return null;
-  return resJson?.domain as string;
-};
